Refactor TensorName to avoid dependency issues
